refactor(app): :recycle: Broke out ble_controller into logical groups (separation of concerns)
